Forsight-backend-staging-main/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.4
billiard==4.2.1
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.21.1
confection==0.1.5
confluent-kafka==2.10.1
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.11
Django==5.2.3
django-cors-headers==4.7.0
django-prometheus==2.4.1
django-storages==1.14.6
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
emoji==2.14.1
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.4.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==6.2.0
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
rich==14.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
tqdm==4.67.1
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
wordcloud==1.9.4
wrapt==1.17.2
zopfli==0.2.3.post1
gunicorn==21.2.0
